refactor: change structure of the code
