+++ src/components/Footer/Footer.jsx
@@ -1,5 +1,5 @@
-import LinkFooter from './LinkFooter';
-import PrivacyFooter from './PrivacyFooter';
+import LinkFooter from './components/LinkFooter';
+import PrivacyFooter from './components/PrivacyFooter';
 
 export default function Footer() {
   return (
