+++ src/components/Footer/components/PrivacyFooter.jsx
