+++ src/components/Home/Shop/components/Action.jsx
@@ -1,0 +1,7 @@
+export const ACTION = {
+    ALL: 'all',
+    WOMEN: `women's clothing`,
+    MEN: `men's clothing`,
+    JEWELERY: 'jewelery',
+    ELECTRONICS: 'electronics'
+};
