+++ src/components/Navbar/components/Menucart/MenuCart.jsx
@@ -1,5 +1,5 @@
-import { useCart, useTotalPriceCart } from '../Provider/CartProvider';
-import CardCart from './CardCart';
+import { useCart, useTotalPriceCart } from '../../../Provider/CartProvider';
+import CardCart from './components/CardCart';
 import { Link } from 'react-router-dom';
 import { PropTypes } from 'prop-types';
